requirements.lock.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
pyyaml==6.0.2
mlflow==2.18.0
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
httpx==0.28.1

black==24.10.0
ruff==0.7.4
pytest==8.3.4

# Strong model (optional but recommended)
xgboost==2.1.3

